requirements.txt
EasyProcess==0.3
entrypoint2==0.2.1
MouseInfo==0.1.3
mss==6.0.0
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
playsound==1.2.2
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pynput==1.6.8
pyperclip==1.8.0
PyRect==0.1.4
pyscreenshot==2.2
PyScreeze==0.1.26
pytesseract==0.3.4
PyTweening==1.0.3
six==1.15.0
